backend/requirements.txt
Flask==3.1.2
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
gunicorn==25.1.0
joblib==1.5.3
numpy==2.4.2
pandas==3.0.1
Pillow==11.3.0
python-dateutil==2.9.0.post0
pytest==8.4.2
pymongo[srv]==4.15.5
python-dotenv==1.2.1
requests==2.32.5
razorpay==1.4.2
scikit-learn==1.8.0
scipy==1.17.0
threadpoolctl==3.6.0
tzdata==2025.3
deepface==0.0.93
opencv-python-headless==4.12.0.88
